Moved newCartographer functions to their own files, scoped them with namespaces, moved some functions to utilities
