+++ AtlasUtil/include/AtlasUtil/Path.h
@@ -1,7 +1,6 @@
 #pragma once
 #include "AtlasUtil/Annotate.h"
 #include "AtlasUtil/Exceptions.h"
-#include "llvm/Analysis/CallGraph.h"
 #include <llvm/IR/BasicBlock.h>
 #include <llvm/IR/CFG.h>
 #include <llvm/IR/Instructions.h>
@@ -352,45 +351,4 @@ static std::set<llvm::BasicBlock *> GetEntrances(std::set<llvm::BasicBlock *> &b
         }
     }
     return Entrances;
-}
-
-static llvm::CallGraph getCallGraph(llvm::Module *mod, std::map<int64_t, int64_t> &blockCallers, std::map<llvm::BasicBlock *, llvm::Function *> &BlockToFPtr, std::map<int64_t, llvm::BasicBlock *> &IDToBlock)
-{
-    llvm::CallGraph CG(*mod);
-    // Add function pointers
-    for (auto &f : *mod)
-    {
-        for (auto bb = f.begin(); bb != f.end(); bb++)
-        {
-            for (auto it = bb->begin(); it != bb->end(); it++)
-            {
-                if (auto CI = llvm::dyn_cast<llvm::CallInst>(it))
-                {
-                    auto callee = CI->getCalledFunction();
-                    if (callee == nullptr)
-                    {
-                        // try to find a block caller entry for this function, if it's not there we have to move on
-                        auto BBID = GetBlockID(llvm::cast<llvm::BasicBlock>(bb));
-                        if (blockCallers.find(BBID) != blockCallers.end())
-                        {
-                            auto calleeID = blockCallers[BBID]; // should only be 1 element long, because of basic block splitting
-                            auto calleeBlock = IDToBlock[calleeID];
-                            if (calleeBlock != nullptr)
-                            {
-                                auto parentNode = CG.getOrInsertFunction(bb->getParent());
-                                auto childNode = CG.getOrInsertFunction(calleeBlock->getParent());
-                                parentNode->addCalledFunction(CI, childNode);
-                                BlockToFPtr[llvm::cast<llvm::BasicBlock>(bb)] = calleeBlock->getParent();
-                            }
-                        }
-                        else
-                        {
-                            spdlog::warn("BlockCallers did not contain an entry for the indirect call in BBID " + std::to_string(BBID));
-                        }
-                    }
-                }
-            }
-        }
-    }
-    return CG;
 }
