--- cartographer/new/inc/newCartographer.h
@@ -1,359 +1,0 @@
-#pragma once
-#include <deque>
-#include <map>
-#include <set>
-#include <vector>
-
-enum class NodeColor
-{
-    White,
-    Grey,
-    Black
-};
-
-template <typename T>
-struct PairComp
-{
-    using is_transparent = void;
-    bool operator()(const std::pair<T, T> &lhs, const std::pair<T, T> &rhs) const
-    {
-        return lhs.first < rhs.first;
-    }
-    bool operator()(const std::pair<T, T> &lhs, T rhs) const
-    {
-        return lhs.first < rhs;
-    }
-    bool operator()(T lhs, const std::pair<T, T> &rhs) const
-    {
-        return lhs < rhs.first;
-    }
-};
-
-struct DijkstraNode
-{
-    DijkstraNode() = default;
-    DijkstraNode(double d, uint64_t id, uint64_t p, NodeColor c)
-    {
-        distance = d;
-        NID = id;
-        predecessor = p;
-        color = c;
-    }
-    uint64_t NID; // maps this dijkstra node to a GraphNode.NID
-    /// distance between this node and the target source node
-    /// since our objective is to find the maximum likelihood path, we need to map probabilities onto a space that minimizes big probabilities and maximizes small ones
-    /// -log(p) is how we do this
-    double distance;
-    /// minimum-distance predecessor of this node
-    uint64_t predecessor;
-    /// whether or not this node has been investigated
-    NodeColor color;
-};
-
-struct DijkstraCompare
-{
-    using is_transparent = void;
-    bool operator()(const DijkstraNode &lhs, const DijkstraNode &rhs) const
-    {
-        return lhs.distance < rhs.distance;
-    }
-    /*bool operator()(const DijkstraNode &lhs, uint64_t rhs) const
-    {
-        return lhs.distance < rhs;
-    }
-    bool operator()(uint64_t lhs, const DijkstraNode &rhs) const
-    {
-        return lhs < rhs.distance;
-    }*/
-} DCompare;
-class GraphNode
-{
-public:
-    uint64_t NID;
-    /// BBIDs from the source bitcode that are represented by this node
-    /// Each key is a member BBID and its value is the basic block its unconditional edge points to
-    /// If a key maps to itself, there is no edge attached to this block
-    std::map<int64_t, int64_t> blocks;
-    /// Maps a neighbor nodeID to a probability edge. The set of keys is comprehensive for all neighbors of this GraphNode
-    /// The first index in the pair is the raw count, the second is the histogram probability
-    std::map<uint64_t, std::pair<uint64_t, double>> neighbors;
-    /// Holds the node IDs of each predecessor of this node
-    std::set<uint64_t> predecessors;
-    GraphNode()
-    {
-        NID = getNextNID();
-        blocks = std::map<int64_t, int64_t>();
-        neighbors = std::map<uint64_t, std::pair<uint64_t, double>>();
-        predecessors = std::set<uint64_t>();
-    }
-    /// Meant to be constructed from a new block description in the input binary file
-    GraphNode(uint64_t ID)
-    {
-        NID = ID;
-        blocks = std::map<int64_t, int64_t>();
-        neighbors = std::map<uint64_t, std::pair<uint64_t, double>>();
-        predecessors = std::set<uint64_t>();
-    }
-    ~GraphNode() = default;
-    void addBlock(int64_t newBlock)
-    {
-        // to add a block
-        // 1.) find the key that maps to itself
-        // 2.) update that value to addBlock
-        // 3.) add a new pair to the map (newBlock,newBlock)
-        for (auto &k : blocks)
-        {
-            if (k.first == k.second)
-            {
-                k.second = newBlock;
-                break;
-            }
-        }
-        blocks[newBlock] = newBlock;
-    }
-    void addBlocks(const std::map<int64_t, int64_t> &newBlocks)
-    {
-        // TODO: finish this
-        for (const auto &k : newBlocks)
-        {
-            blocks[k.first] = k.first;
-        }
-        // to add this tree of blocks
-        // 1.) find the block that ends the current node
-        // 2.) connect that block to the end block
-        // 3.) append the rest of the new blocks to the map
-    }
-    virtual void addNodeBlocks(GraphNode *newNode){};
-
-protected:
-    static uint64_t nextNID;
-    static uint64_t getNextNID()
-    {
-        return nextNID++;
-    }
-};
-
-/// Allows for us to search a set of GraphNodes using an NID
-struct p_GNCompare
-{
-    using is_transparent = void;
-    bool operator()(const GraphNode *lhs, const GraphNode *rhs) const
-    {
-        return lhs->NID < rhs->NID;
-    }
-    bool operator()(const GraphNode *lhs, uint64_t rhs) const
-    {
-        return lhs->NID < rhs;
-    }
-    bool operator()(uint64_t lhs, const GraphNode *rhs) const
-    {
-        return lhs < rhs->NID;
-    }
-};
-struct GNCompare
-{
-    using is_transparent = void;
-    bool operator()(const GraphNode &lhs, const GraphNode &rhs) const
-    {
-        return lhs.NID < rhs.NID;
-    }
-    bool operator()(const GraphNode &lhs, uint64_t rhs) const
-    {
-        return lhs.NID < rhs;
-    }
-    bool operator()(uint64_t lhs, const GraphNode &rhs) const
-    {
-        return lhs < rhs.NID;
-    }
-};
-/// @brief Dijkstra implementation inspired by boost library
-std::vector<uint64_t> Dijkstras(const std::set<GraphNode, GNCompare> &nodes, uint64_t source, uint64_t sink);
-
-/// @brief Virtual Kernel Node
-///
-/// Represents a virtualized kernel. A virtualized kernel node is a node that represents an entire kernel, and all nodes that once were within the subgraph itself are now contained within this structure
-class VKNode : public GraphNode
-{
-public:
-    struct Kernel *kernel;
-    VKNode(struct Kernel *p_k) : GraphNode()
-    {
-        kernel = p_k;
-    }
-    VKNode(const GraphNode &GN, struct Kernel *p_k)
-    {
-        NID = GN.NID;
-        kernel = p_k;
-        blocks = GN.blocks;
-        neighbors = GN.neighbors;
-        predecessors = GN.predecessors;
-    }
-    ~VKNode() = default;
-    void addNodeBlocks(GraphNode *newNode)
-    {
-        for (const auto &block : newNode->blocks)
-        {
-            blocks[block.first] = block.second;
-        }
-    }
-};
-
-struct Kernel
-{
-    std::set<GraphNode, GNCompare> nodes;
-    VKNode *kernelNode;
-    uint32_t KID;
-    Kernel()
-    {
-        KID = getNextKID();
-        nodes = std::set<GraphNode, GNCompare>();
-    }
-    Kernel(uint32_t ID)
-    {
-        KID = ID;
-        nodes = std::set<GraphNode, GNCompare>();
-    }
-    /// Returns the IDs of the kernel entrances
-    /// @retval    entrances Vector of IDs that specify which nodes (or blocks) are the kernel entrances. Kernel entrances are the source nodes of edges that enter the kernel
-    std::vector<GraphNode> getEntrances() const
-    {
-        std::vector<GraphNode> entrances;
-        for (const auto &node : nodes)
-        {
-            for (const auto &pred : node.predecessors)
-            {
-                if (nodes.find(pred) == nodes.end())
-                {
-                    // this node has a predecessor outside of the kernel, so it is considered an entrance node
-                    entrances.push_back(node);
-                    break;
-                }
-            }
-        }
-        return entrances;
-    }
-    /// @brief Returns the IDs of the kernel exits
-    ///
-    /// @param[in] allNodes Set of all nodes in the control flow graph. Used to copy the nodes that are the destinations of edges that leave the kernel
-    /// @retval    exits Vector of IDs that specify which nodes (or blocks) are the kernel exits. Kernel exits are nodes that have a neighbor outside the kernel
-    std::vector<GraphNode> getExits() const
-    {
-        std::vector<GraphNode> exitNodes;
-        for (const auto node : nodes)
-        {
-            for (const auto &neighbor : node.neighbors)
-            {
-                if (nodes.find(neighbor.first) == nodes.end())
-                {
-                    // we've found an exit
-                    exitNodes.push_back(node);
-                }
-            }
-        }
-        return exitNodes;
-    }
-    /// @brief Returns the member blocks (from the source bitcode) of this kernel
-    const std::set<int64_t> getBlocks() const
-    {
-        std::set<int64_t> blocks;
-        for (const auto &node : nodes)
-        {
-            for (const auto &block : node.blocks)
-            {
-                blocks.insert(block.first);
-            }
-        }
-        return blocks;
-    }
-    /// @brief Compares this kernel to another kernel by measuring node differences
-    ///
-    /// If two kernels are the same, 1 will be returned
-    /// If two kernels are completely different, 0 will be returned
-    /// If two kernels share some nodes, (compare shared) / (this size) will be returned
-    /// TODO: if this object fully overlaps with compare, but compare contains other blocks, this will say that we fully match when we actually don't. Fix that
-    std::set<int64_t> Compare(const Kernel &compare) const
-    {
-        std::set<int64_t> shared;
-        for (const auto &compNode : compare.nodes)
-        {
-            if (nodes.find(compNode) != nodes.end())
-            {
-                for (const auto &block : compNode.blocks)
-                {
-                    shared.insert(block.first);
-                }
-            }
-        }
-        return shared;
-    }
-    /// Returns true if any node in the kernel can reach every other node in the kernel. False otherwise
-    bool FullyConnected() const
-    {
-        for (const auto &node : nodes)
-        {
-            // keeps track of which nodeIDs have been visited, all initialized to white
-            std::map<uint64_t, NodeColor> colors;
-            for (const auto &node2 : nodes)
-            {
-                colors[node2.NID] = NodeColor::White;
-            }
-            // holds newly discovered nodes
-            std::deque<GraphNode> Q;
-            Q.push_back(node);
-            while (!Q.empty())
-            {
-                for (const auto &neighbor : Q.front().neighbors)
-                {
-                    // check if this neighbor is within the kernel
-                    if (nodes.find(neighbor.first) != nodes.end())
-                    {
-                        if (colors[neighbor.first] == NodeColor::White)
-                        {
-                            Q.push_back(*(nodes.find(neighbor.first)));
-                        }
-                    }
-                    colors[neighbor.first] = NodeColor::Black;
-                }
-                Q.pop_front();
-            }
-            // if any nodes in the kernel have not been touched, this node couldn't reach them
-            for (const auto &node : colors)
-            {
-                if (node.second == NodeColor::White)
-                {
-                    return false;
-                }
-            }
-        }
-        return true;
-    }
-    inline bool operator==(const Kernel &rhs) const
-    {
-        return rhs.KID == KID;
-    }
-
-private:
-    static uint32_t nextKID;
-    static uint32_t getNextKID()
-    {
-        return nextKID++;
-    }
-};
-
-/// Allows for us to search a set of GraphNodes using an NID
-struct KCompare
-{
-    using is_transparent = void;
-    bool operator()(const Kernel *lhs, const Kernel *rhs) const
-    {
-        return lhs->KID < rhs->KID;
-    }
-    bool operator()(const Kernel *lhs, uint64_t rhs) const
-    {
-        return lhs->KID < rhs;
-    }
-    bool operator()(uint64_t lhs, const Kernel *rhs) const
-    {
-        return lhs < rhs->KID;
-    }
-};
